fix type error in user.java

+++ src/main/java/com/ghostchu/sapling/domain/model/User.java
@@ -119,7 +119,7 @@ public class User {
         return userId;
     }
 
-    public int getGroupId() {
+    public long getGroupId() {
         return groupId;
     }
